Parametrize the values for the used adapters
Parametrizing this, we allow to configure automatically which
adapters to use, specifying the class they are implemented.
This helps the definition and automatic deployment on continous
integration systems, such as Jenkins. We were using the default
simple evaluator before, so in the case of Fiware, the metrics
were not properly evaluated.

The change involves parametrizing the folloing parameters:

 * enforcement.constraintEvaluator.class
 * enforcement.metricsRetriever.class