Add the new parameters in the autoconfiguration script
In the previous commit, only the configuration file was parametrized.
When adding new parameters in that file, it is also needed to update
the autoconfiguration script.

In the configuration.properties.sample class, the parameters for the
enforcement.constraintEvaluator.class and the
enforcement.metricsRetriever.class were parametrized, but the sed
commands in the autoconfigure.sh script were not updated to get their
values from the global variables.